提交人：皮扬帆 更新内容： 1.完成了帖子和评论的相关业务

+++ nexus-backend/src/main/java/com/hcz/nexusbackend/controller/CommentController.java
@@ -19,19 +19,19 @@ public class CommentController {
     private CommentService commentService;
 
     @PostMapping("/create")
-    public Comment create(@RequestParam Integer postId,
-                          @RequestParam Integer userId,
+    public Comment create(@RequestParam Long postId,
+                          @RequestParam Long userId,
                           @RequestParam String content) {
         return commentService.create(postId, userId, content);
     }
 
     @GetMapping("/list")
-    public List<Comment> list(@RequestParam Integer postId) {
+    public List<Comment> list(@RequestParam Long postId) {
         return commentService.list(postId);
     }
 
     @PostMapping("/accept")
-    public String accept(@RequestParam Integer commentId) {
+    public String accept(@RequestParam Long commentId) {
         return commentService.accept(commentId);
     }
 }
